matcher/requirements.txt
# fastapi v
# uvicorn v
# pydantic v
# python-multipart v
# opencv-python-headless v
# ultralytics
# scikit-learn 
# googlemaps 
# gmaps
# polyline
# transformers
# openai
# pandas


# --- API / server ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0

# # --- CV / Ultralytics YOLO ---
# opencv-python-headless==4.8.1.78

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
scipy==1.11.4

sentence-transformers==2.6.1
transformers==4.41.2
huggingface_hub==0.23.2

pinecone-client==3.2.2
# pinecone-client==2.2.4
schedule==1.2.1
mysql-connector-python==8.3.0
